007_noise: relative power in gamma range (50-100Hz) and max. frequency
simulation_bump_fitting now Fourier transforms the inhibitory synaptic currents
and extracts relative amount of power in the gamma range (set to 50-100 Hz) and
find the frequency with max. power in this range.

Currently FFT-ing the signal creates lots of theta (8 Hz) harmonics in the power
spectrum, which are distributed in the gamma range; so FFT might not be the best
choice here.