data_storage: interfaces, HDF5 implementation.
The data storage package allows the user to store (almost) arbitrary data in a
dictionary-like format and in a transparent way. It is similar to a shelve
module, but does not necessarily do the pickle operation (why should it btw?)

This is the initial commit, and the HDF5 implementation has been tested in a
basic way. No guarantee that it is bug-free.