data_storage: FIX: order of items maintained when storing lists.
The solution is to store list item names (indexes) as zero-padded strings. At
least for now. Not the best solution as it is not clear whether HDF5 and h5py
guarantee to read the keys in lexicographical order.

Tests updated for checking a moderately item order for a moderately large list
(~5000 int items)