grids_no_velocity simulations
Submit simulations in which the velocity input is disabled, but the place cell
input is enabled. This has to be compared to the full simulations with the same
range of place cell firing rates (pc_conn_weight == 0.5 nS).

THe expected results should be that there is a region of pc_max_rate parameter
in which grid fields are maintained with velocity input, but disappear when
velocity input is switched off, i.e. place cell input is not enough to drive
the network. Otherwise the place cell input is on its own strong enough to
drive the bump.