grids_init simulations, velocity ON (see whole log msg)
This is to test the final version of place cell input together with velocity
integration.
Place cells are active in the beginning, with a very strong input (used in the
previous param. sweep simulations). After the theta input starts, place cells
are active as well, but with much lower connections weights and possibly firing
rates.

This simulations has the velocity input enabled.