full grid field simulation parameter sweeps (150 pA)
These simulations test whether the presence/absence of noise helps in the path
integration process of the network. There is a simulation of the full animal
movement (here in particular 600s), with the initialisation place cells and
velocity place cells, currently for one trial. This trial will be evaluated and
subsequently more trials will be run.

- Saving full population of E and spikes of the first 10 I neurons
- Every simulation should have its bumpCurrentSlope parameter correctly set up
  from the estimated data
- if bumpCurrentSlope does not exist in a simulation, it is aborted
- A -notify option added to qsub and when the simulation receives a signal
  (SIGUSR2) it attempts to store all the data before SIGKILL arrives