CHANGE: BumpVelocityVisitor maximizes the fits to <0, bumpSpeedMax>
Determining the best fit of the bump speed vs. injected current relationship
============================================================================
It is necessary to set the correct range of IvelVec in order to obtain the best
fit.
This parameter sets the desired range of _bump_ speeds that the line fit should
cover. In case the range is not covered within the available values of Ivel,
the fit with the whole range will be used.

This has been tested on a few data points, works
TODO: Instead of picking the whole IvelVec range, pick the fit with the maximal
range of bump speeds.