UPDATE submitting pkg: CPU-based non-blocking submission of jobs
This update allows to submit up to numCPU processes concurrently. numCPU is not
related to the number of processors and the system will not attempt to detect
whether it is more than the cpu count on the machine.

Setting blocking to False has the original effect: the numCPU is ignored (might
be removed or changed in the future).

This has been tested on WorkstationSubmitter class only.