refactor place cell settings
A user can now provide the posIn argument for
NestGridCellNetwork.setPlaceCells() which will force the place cell input to
follow the path specified in posIn. If posIn is None, then the standard way of
place cell definition will be performed: loaded from the rat velocity file and
positional input will then be obtained from that file.