Change velocity simulations to VERTICAL bump movement
This changes the output directory to 'even_spacing/velocity_vertical'.

This has been accomplished simply by changing the const_v parameter in
simulation_velocity.py

In previous versions, the bump movement was horizontal. This caused problems in
position estimation, because if the bump started close to a vertical boundary
the X position could in effect flicker during the simulation and disrupt the
estimation of bump positions.