reopen #1: modulo boundaries introduced bug to network generation
In short: when generating distances to compute synapses on the twisted torus,
some of the values were shifted. This introduced problems, because the shift
was not implemented on a twisted torus.

Solution to be tested: respect real twisted-toroidal boundaries when doing the
shifts in GridCellNetwork._generateGaussianWeights().

This seems to produce *almost* identical connections weights to those generated
from the earlier revisions. However the earlier generation was still incorrect,
due to not wrapping around the shifted value.