add gaussian fitting unit test
This test is not ideal, since in some cases it fails to estimate the parameters
properly. This is most likely due to the nature of the twisted torus, which can
be ambiguous. It is thus possible that in some cases this ambiguity messes up
with the gradient of the least squares method.

Currently, the test passes if there are less than 10% incorrect estimations.