implement the E-->E connections (does not produce bump yet)
This is a prototype of E-->E connections. However it was not yet possible to
make the network spontaneously form a bump state. Perhaps it is not even
possible at all, since the network might behave in a bistable/multistable way,
cf. Camperi et al., 1998; Compte et al., 2000, or perhaps even Roxin et al.,
2005.